+++ src/components/Layout/RightNavigation.js
@@ -1,11 +1,11 @@
 import React from 'react';
 
 const RightNavigation = () => {
-  const title = 'Right panel';
+  const content = 'Put more content here';
 
   return (
     <div className="flex flex-col w-24 text-gray-700 min-h-screen justify-center">
-      { title }
+      { content }
     </div>
   );
 };