+++ src/components/shared/__tests_/CheckBox.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import CheckBox from '../CheckBox';
+
+describe('test CheckBox component', () => {
+  it('should render without crashing', () => {
+    const wrapper = shallow(<CheckBox />);
+
+    expect(wrapper).toMatchInlineSnapshot('ShallowWrapper {}');
+  });
+});